Fix Vercel function count for storefront checkout
