+++ api/stripe/[route].js
@@ -3,6 +3,16 @@ import { resolvePlanFromSubscription } from '../../src/money/access.js';
 import { createStripeCheckoutHandler } from '../../src/billing/api-checkout.js';
 import { createStripeStatusHandler } from '../../src/billing/api-status.js';
 
+const SAMPLE_STOREFRONT_PRODUCTS = Object.freeze({
+  'compact-desk-dock': Object.freeze({
+    name: 'Compact Desk Dock',
+    description: 'A sample dropship product page for testing single-product demand.',
+    unitAmount: 7900,
+    currency: 'usd',
+    image: 'https://images.unsplash.com/photo-1516321318423-f06f85e504b3?auto=format&fit=crop&w=1200&q=80',
+  }),
+});
+
 function makeStripeClient(config = process.env) {
   const secretKey = String(config?.STRIPE_SECRET_KEY || '').trim();
   if (!secretKey) {
@@ -624,6 +634,105 @@ function getRouteValue(req) {
   return String(route || '').trim().toLowerCase();
 }
 
+function firstHeaderValue(value = '') {
+  return String(value || '').split(',')[0].trim();
+}
+
+function getRequestOrigin(req, config = process.env) {
+  const host = firstHeaderValue(req?.headers?.['x-forwarded-host'] || req?.headers?.host || '');
+  if (host) {
+    const proto = firstHeaderValue(req?.headers?.['x-forwarded-proto'] || '').toLowerCase()
+      || (host.startsWith('localhost') || host.startsWith('127.0.0.1') ? 'http' : 'https');
+    return `${proto}://${host}`;
+  }
+  return String(config?.PORTAL_ORIGIN || 'https://portal.3dvr.tech').trim().replace(/\/+$/, '');
+}
+
+function readBody(req) {
+  return req?.body && typeof req.body === 'object' ? req.body : {};
+}
+
+function normalizeStorefrontQuantity(value) {
+  const quantity = Number.parseInt(value, 10);
+  if (!Number.isFinite(quantity)) return 1;
+  return Math.min(Math.max(quantity, 1), 5);
+}
+
+async function createStorefrontCheckoutSession(req, stripeClient, res, config = process.env) {
+  if (req.method === 'OPTIONS') {
+    return res.status(200).end();
+  }
+
+  if (req.method === 'GET') {
+    return res.status(200).json({
+      stripeConfigured: Boolean(config.STRIPE_SECRET_KEY),
+      products: Object.keys(SAMPLE_STOREFRONT_PRODUCTS),
+    });
+  }
+
+  if (req.method !== 'POST') {
+    res.setHeader('Allow', 'GET, POST, OPTIONS');
+    return res.status(405).json({ error: 'Method Not Allowed' });
+  }
+
+  const body = readBody(req);
+  const productId = String(body.productId || '').trim();
+  const product = SAMPLE_STOREFRONT_PRODUCTS[productId];
+  const quantity = normalizeStorefrontQuantity(body.quantity);
+  const orderId = String(body.orderId || '').trim().slice(0, 80);
+  const customerEmail = String(body.customerEmail || '').trim().toLowerCase();
+  const customerName = String(body.customerName || '').trim().slice(0, 120);
+  const origin = getRequestOrigin(req, config);
+
+  if (!product) {
+    return res.status(400).json({ error: 'Unknown product.' });
+  }
+
+  if (!orderId) {
+    return res.status(400).json({ error: 'Missing order id.' });
+  }
+
+  const session = await stripeClient.checkout.sessions.create({
+    mode: 'payment',
+    success_url: `${origin}/victor-dropship/?checkout=success&order=${encodeURIComponent(orderId)}&session_id={CHECKOUT_SESSION_ID}`,
+    cancel_url: `${origin}/victor-dropship/?checkout=cancel&order=${encodeURIComponent(orderId)}`,
+    customer_email: customerEmail || undefined,
+    billing_address_collection: 'auto',
+    shipping_address_collection: {
+      allowed_countries: ['US'],
+    },
+    metadata: {
+      storefront: 'victor-dropship-sample',
+      order_id: orderId,
+      product_id: productId,
+      customer_name: customerName,
+    },
+    line_items: [
+      {
+        quantity,
+        price_data: {
+          currency: product.currency,
+          unit_amount: product.unitAmount,
+          product_data: {
+            name: product.name,
+            description: product.description,
+            images: [product.image],
+            metadata: {
+              product_id: productId,
+              fulfillment: 'manual-vendor-order',
+            },
+          },
+        },
+      },
+    ],
+  });
+
+  return res.status(200).json({
+    id: session.id,
+    url: session.url,
+  });
+}
+
 async function listCustomers(stripeClient, res) {
   const invoicesIterator = stripeClient.invoices.list({
     status: 'paid',
@@ -739,6 +848,14 @@ export function createStripeDashboardHandler({
     }
     if (route === 'status') {
       return statusHandler(req, res);
+    }
+    if (route === 'storefront-checkout') {
+      try {
+        return await createStorefrontCheckoutSession(req, resolvedStripeClient, res, config);
+      } catch (error) {
+        console.error('Failed to create Stripe storefront checkout', error);
+        return res.status(500).json({ error: 'Unable to create Stripe storefront checkout.' });
+      }
     }
 
     if (!resolvedStripeClient) {
